PHANTHOM SECURITY PROJECT/requirements.txt
# Core Dependencies
Flask>=2.3.0
requests>=2.31.0
pyyaml>=6.0

# Machine Learning
scikit-learn>=1.3.0
numpy>=1.24.0

# Database and Caching
redis>=4.5.0

# GeoIP
geoip2>=4.7.0

# Security
cryptography>=41.0.0
python-jose>=3.3.0

# Monitoring and Logging
prometheus-client>=0.17.0
python-json-logger>=2.0.7

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Web Frameworks (Optional - for integrations)
django>=4.2.0
fastapi>=0.100.0
uvicorn>=0.23.0

# Utilities
python-dateutil>=2.8.2
validators>=0.20.0
user-agents>=2.2.0
